Show the ball's y coordinate on the minimap.
y is up and down on the big display.
z is up and down on the mini display because that is top down.
y would be in and out of the sceen to make the mini display consistent.

So I make the ball bigger when it is closer to the viewer on the mini display.
And smaller when it is further away.

Thsi is similar to what the big display is doing.
But I'm using a much simpler algorithm.
And it works well!

This also solves another problem.
I couldnt decide between 1 and 3 for the radius of the mini ball.
Now I'm using the entire range between 1 and 3!