Several improvements to the GUI.
Remove the internal grid lines.
Add grid lines along the back wall.

Disable the drawLines() test.

Add drawText() which doesn't seem to do anything.  🙁

Draw the walls in different colors.
Done in drawPlanes().
This is simple but it looks a lot better.

Speed up the ball.
The initial speed was too slow.
It looks better when it moves quickly.

Add a help hyperlink.
Click on the ? to go to the GitHub page with README.md.