Import the font asset the correct way.
The font only worked right when running locally.
Now it works when I build / publish.

The text is still hiding.
See the new TODO item for new details.

Import a *.json file and get the text from there.
It was hard coded to "Pow!!".
Now the program knows a variety of Batman phrases.