Code reorganization:  Use CanvasTexture
I was using a more complicated method
to copy the contents of the texture into a texture.

The new method gets rid of a callback and a promise.
I can do the update in an animationFrame callback now.
And when I'm ready to do more updates, that will be easy.