Make the wall wave when it is struck.
This is the best the demo has looked in a while!

I temporarily disabled the batman fight words
because theyw were getting in the way of the new effect.
The plan is to display one or the other, chosen somwhat randomly.

Remove the constantly pulsing walls.
The new effect is a variation on that effect.
Remove the extra canvas that showed the displacementMap.
That was only there for debug reasons.