Only show one special effect at a time.
tic-tac-toe has the highest priority.
If that move is unavailable, use the BumpEffect.

Lots of automatic reformattng.
More than I would have expected!
Some is my fault for not formatting the code each commit.
But now there are blank lines
above and below the <head> and <body> sections.
Did the rules change on me?