Improve the tic-tac-toe drawings.
If you win the tic-tac-toe game, we draw a line to show it.
This commit makes the line thicker and randomer.

Lots of Internal cleanup.

Add a lot of documentation.
Mostly to the TicTacToe class.

Remove some old comments.
And unused variables.
And old log statements.

Notice the new TODO/bug.
This is a new report, not a new bug.