Add bullet holes.
This is just the beginning for the bullet hole special effect.
Ideally these would be mixed with the 3d graphics.
The program will add these as needed.

Fix the tic-tac-toe bug.
I fixed the source code in the previous commit.
But I forgot to do a build before making that commit.