Add the broken glass effect.
If you hit the front wall, you see a crack.
The bulk of the front wall is still transparent.

The other walls didn't change.
I had to reorganize the code a lot so the front wall could share code.
But none of these affect the functionality of the walls.

I cropped glass.png.
It's square now.
And the hole is in the center.
That made the code simpler.