Add a star effect to the front wall.
Sometimes, when the ball bounces off the front wall,
the program draws a star.
Or draw the cracked glass.
It's a random choice on each bounce.

This is the same star that was previously displayed on the other walls
I plan to bring it back.
Like on the front wall,
there should be a random chance of seeing a star.