Draw more shapes on the front wall.
The program can now draw and X or an O.
Use the right positions for a complete  tic-tac-toe board,
even though it's just a single X or O and no board.
Randomly pick between X, O, star, and cracked glass.

The front wall is getting close to my plans for all walls.
There will be lots of shapes and things to draw.
The program will pick one somewhat randomly.

I'm using the exact same code to draw X or O on any wall.
The other walls still look the same.
I moved that code so it could be shared,
but otherwise that code did not change.