Add more effects to the walls.
Add the Batman fight words back.
These are similar to what we had before.
They now accept shadows.
There is some randomness to the hight, now.
The back wall now adds a bevel.

Draw a star on the side and back walls sometimes.
This is the same star that the code already drew on the front wall.

Randomly pick from the available actions.