The program sometimes clears what's one a wall.
Sometimes the program will erase everything
and draw a fresh tic tac toe board.
And reset the tic tac toe game.
After a brief animation, this retores the initial state of the wall.

Bug fix:
Sometimes the program would try to make a tic tac toe move
in a cell that was already full.
The effect was to do nothing.
Now we just pick a different effect, one that works.

Remove a TODO item.
There is a now small chance of drawing a star
over an active tic tac toe game.
That will end the tic tac toe game.