First serious attempt at the bullet hole effect.
I removed the bullet hole that was always present.
That was a quick prototype
and a reminder to do it right.
Now everything that is on the screen looks decent.

The bullet hole effect is only available via the JavaScript console.
Check out random-notes.txt for more details.

I'm not completely happy with the way I'm handling the aspect ratio.
I'm committing here because I've made some good progress.
I want to clean that code up a little,
before the main program tires to draw this effect.